Update buildPackage.py
Encapsulate the Program to manage the source code packaged or debugged by Dev. The buildPackage tool uses programPool to call the relevant source code for other operations, which is considered to improve the code structure and operation logic of the original part. I left a TODO, I won't write it for now, I will improve it later.